Sync our tree with upstream official source patches from 2015-2017, inclusive.

+++ ax25tools/tcpip/rip98d.h
@@ -48,8 +48,8 @@ extern int route_restrict;
 extern int logging;
 
 /* In rip98d.c */
-extern unsigned int mask2bits(unsigned long int);
-extern unsigned long int bits2mask(unsigned int);
+extern unsigned int mask2bits(unsigned int);
+extern unsigned int bits2mask(unsigned int);
 
 /* In rip98t.c */
 extern void transmit_routes(int);
